ResultBearing: split out ResultIterable
